+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/api/MagikPunctuator.java
@@ -21,6 +21,21 @@ public enum MagikPunctuator implements GrammarRuleKey {
 
   private final String value;
 
+  /**
+   * Get all punctuator values.
+   *
+   * @return Punctuator values
+   */
+  public static String[] punctuatorValues() {
+    final String[] punctuatorValuess = new String[MagikPunctuator.values().length];
+    int idx = 0;
+    for (final MagikPunctuator punctuator : MagikPunctuator.values()) {
+      punctuatorValuess[idx] = punctuator.getValue();
+      idx++;
+    }
+    return punctuatorValuess;
+  }
+
   MagikPunctuator(final String value) {
     this.value = value;
   }