+++ CHANGES.md
@@ -36,6 +36,7 @@
 - Fix auto formatting bug where `_pragma` caused lines to be removed.
 - FormattingCheck now uses formatting code to detect issues.
 - Formatter now supports ranged formatting.
+- Add `magik-session-wrapper` to wrap a Magik running image/session, providing a better CLI experience.
 - Several fixes.
 
 ### Breaking changes (reiterated from above)